cleaned and updated back end and front end code

+++ controllers/pages/delete.js
@@ -1,10 +1,17 @@
-const Page = require('../../database/models/Pages');
+const Page = require("../../database/models/Pages");
 
 module.exports = (req, res) => {
-  Page.findByIdAndDelete(req.params.id, (err, doc) => {
-    if(!err) {
-      res.redirect('/admin');
-    } else {res.redirect('/login')}
+  const authenticated = req.isAuthenticated();
 
-  });
-}
+  if (authenticated) {
+    Page.findByIdAndDelete(req.params.id, (err, doc) => {
+      if (!err) {
+        res.redirect("/admin");
+      } else {
+        res.redirect("/login");
+      }
+    });
+  } else {
+    res.redirect("/login");
+  }
+};
